requirements.txt
lxml==6.0.2
pillow==12.1.1
python-pptx==1.0.2
typing_extensions==4.15.0
xlsxwriter==3.2.9
requests==2.32.3
beautifulsoup4==4.12.3
